Extract ForkContext into separate file; add a current_fork field
